% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnualGraph.R
\name{annualGraph}
\alias{annualGraph}
\title{annual graph
annual graph in treatment regimen}
\usage{
annualGraph(connectionDetails, resultDatabaseSchema, cohortTable,
  targetCohortIds, conditionCohortIds = NULL, fromYear, toYear)
}
\value{
annual treatment regimen highcharter graph
}
\description{
annual graph
annual graph in treatment regimen
}
\keyword{year}
