% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Visualization.R
\name{CancerTxPatterns}
\alias{CancerTxPatterns}
\title{sankey for regimen and surgery
Visualization tool for sankey for regimen and surgery}
\usage{
CancerTxPatterns(connectionDetails, oracleTempSchema, cdmDatabaseSchema,
  cohortDatabaseSchema, oncologyDatabaseSchema, vocaDatabaseSchema,
  cohortTable, episodeTable, outputFolder, outputFileTitle,
  targetCohortIds, episodeCohortCreate = FALSE, createEpisodeCohortTable,
  fromYear = 1998, toYear = 2018, identicalSeriesCriteria = 60,
  maximumCycleNumber = 18, minSubject = 0, collapseDates = 0,
  conditionCohortIds, treatmentLine = 3, minimumRegimenChange = 1,
  surgeryCohortIds, eventCohortIds)
}
\value{
CancerTxPatterns plots
}
\description{
sankey for regimen and surgery
Visualization tool for sankey for regimen and surgery
}
