% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterationHeatmap.R
\name{cohortCycle}
\alias{cohortCycle}
\title{Heatmap
Trends in regimen heatmap}
\usage{
cohortCycle(connectionDetails, resultDatabaseSchema, cohortTable,
  targetCohortIds, identicalSeriesCriteria, conditionCohortIds)
}
\value{
repitition trend heatmap
}
\description{
Heatmap
Trends in regimen heatmap
}
\keyword{heatmap}
