% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullCohort.R
\name{cohortRecords}
\alias{cohortRecords}
\title{CallRawData
Bring the chemotherapy raw data table.}
\usage{
cohortRecords(connectionDetails, resultDatabaseSchema, cohortTable,
  targetCohortIds)
}
\value{
raw data for visualization
}
\description{
CallRawData
Bring the chemotherapy raw data table.
}
\keyword{data}
\keyword{raw}
