% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohort.R, R/CreateEpisodeCohort.R
\name{createCohort}
\alias{createCohort}
\title{CreateCohort
Create cohort of interest by using condition concept Ids.}
\usage{
createCohort(createCohortTable = F, connectionDetails,
  oracleTempSchema = NULL, cdmDatabaseSchema, cohortDatabaseSchema,
  oncologyDatabaseSchema, vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortTable, episodeTable, conceptIdSet = c(),
  includeConceptIdSetDescendant = F, collapseGapSize = 0,
  targetCohortId, cycle = TRUE)

createCohort(createCohortTable = F, connectionDetails,
  oracleTempSchema = NULL, cdmDatabaseSchema, cohortDatabaseSchema,
  oncologyDatabaseSchema, vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortTable, episodeTable, conceptIdSet = c(),
  includeConceptIdSetDescendant = F, collapseGapSize = 0,
  targetCohortId, cycle = TRUE)
}
\description{
CreateCohort
Create cohort of interest by using condition concept Ids.

CreateCohort
Create cohort of interest by using condition concept Ids.
}
