% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEpisodeCohort.R
\name{createEpisodeCohort}
\alias{createEpisodeCohort}
\title{createEpisodeCohort
Create cohort of interest by using condition concept Ids.}
\usage{
createEpisodeCohort(connectionDetails, oracleTempSchema = NULL,
  cdmDatabaseSchema, cohortDatabaseSchema, oncologyDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema, cohortTable,
  episodeTable, conceptIdSet = c(), includeConceptIdSetDescendant = F,
  collapseGapSize = 0, targetCohortId, cycle = TRUE)
}
\description{
createEpisodeCohort
Create cohort of interest by using condition concept Ids.
}
