% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CycleIncidencePlot.R
\name{cycleIncidencePlot}
\alias{cycleIncidencePlot}
\title{Incidence comparison plot
Compare the incidence of event in each treatment}
\usage{
cycleIncidencePlot(connectionDetails, cohortDatabaseSchema, cohortTable,
  outputFolder = NULL, outputFileTitle = NULL, targetCohortIds,
  conditionCohortIds, eventCohortIds, restrictInitialSeries = TRUE,
  restricInitialEvent = TRUE, identicalSeriesCriteria = 60,
  eventPeriod = 30, targetMin = 0)
}
\value{
Incidence plot
}
\description{
Incidence comparison plot
Compare the incidence of event in each treatment
}
\keyword{Incidence}
