% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentIterationHeatmap.R
\name{distributionTable}
\alias{distributionTable}
\title{Heatmap
Trends in regimen heatmap}
\usage{
distributionTable(standardData, targetId)
}
\value{
repitition trend heatmap
}
\description{
Heatmap
Trends in regimen heatmap
}
\keyword{heatmap}
