% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_RegimenExtraction.R
\name{executeExtraction}
\alias{executeExtraction}
\title{Main}
\usage{
executeExtraction(connectionDetails, oracleTempSchema = NULL,
  cdmDatabaseSchema, cohortDatabaseSchema,
  vocaDatabaseSchema = cdmDatabaseSchema,
  oncologyDatabaseSchema = cdmDatabaseSchema,
  createCohortTable = FALSE, createEpisodeAndEventTable = FALSE,
  episodeTable, episodeEventTable, cohortTable, maxCores = 4)
}
\value{
Episode table, Episode Event table, cancer cohort
}
\description{
Main
}
