% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceDatePlot.R
\name{incidenceDatePlot}
\alias{incidenceDatePlot}
\title{Violin Plot
First incidence of event}
\usage{
incidenceDatePlot(connectionDetails, cohortDatabaseSchema, cohortTable,
  targetCohortIds, outputFolder = NULL, outputFileTitle = NULL,
  identicalSeriesCriteria, conditionCohortIds, eventCohortIds,
  restrictEventDate)
}
\value{
Incidence plot
}
\description{
Violin Plot
First incidence of event
}
\keyword{Incidence}
