% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidencePlot.R
\name{incidencePlot}
\alias{incidencePlot}
\title{Incidence comparison plot
Compare the incidence of event in each treatment}
\usage{
incidencePlot(connectionDetails, resultDatabaseSchema, cohortTable,
  targetCohortIds, conditionCohortIds, eventCohortIds,
  restrictInitialSeries = TRUE, restricInitialEvent = TRUE,
  identicalSeriesCriteria = 60, eventPeriod = 30, targetMin = 0)
}
\value{
Incidence plot
}
\description{
Incidence comparison plot
Compare the incidence of event in each treatment
}
\keyword{Incidence}
