% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPathway.R
\name{treatmentPathway}
\alias{treatmentPathway}
\title{sankey for regimen and surgery
Visualization tool for sankey for regimen and surgery}
\usage{
treatmentPathway(connectionDetails, cohortDatabaseSchema, cohortTable,
  outputFolder = NULL, outputFileTitle = NULL,
  conditionCohortIds = NULL, targetCohortIds, eventCohortIds = NULL,
  minimumRegimenChange = 1, treatmentLine = 3, collapseDates = 0,
  minSubject = 0)
}
\value{
sankey for regimen with other evnets
}
\description{
sankey for regimen and surgery
Visualization tool for sankey for regimen and surgery
}
\keyword{sankey}
