% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsagePatternGraph.R
\name{usagePatternGraph}
\alias{usagePatternGraph}
\title{usagePattern graph
usagePattern graph in treatment regimen}
\usage{
usagePatternGraph(connectionDetails, cohortDatabaseSchema, cohortTable,
  targetCohortIds, conditionCohortIds = NULL, outputFolder = NULL,
  outputFileTitle = NULL, fromYear, toYear)
}
\value{
usagePattern treatment regimen highcharter graph
}
\description{
usagePattern graph
usagePattern graph in treatment regimen
}
\keyword{year}
