% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViolinPlot.R
\name{violinPlot}
\alias{violinPlot}
\title{Violin Plot
First incidence of event}
\usage{
violinPlot(connectionDetails, resultDatabaseSchema, cohortTable,
  targetCohortIds, identicalSeriesCriteria, conditionCohortIds,
  eventCohortIds, restrictEventDate, restrictInitialSeries)
}
\value{
Incidence plot
}
\description{
Violin Plot
First incidence of event
}
\keyword{Incidence}
