% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohort.R
\name{createCohort}
\alias{createCohort}
\title{CreateCohort
Create cohort of interest by using condition concept Ids.}
\usage{
createCohort(createCohortTable = F, connectionDetails,
  oracleTempSchema = NULL, cdmDatabaseSchema, cohortDatabaseSchema,
  vocaDatabaseSchema = cdmDatabaseSchema, cohortTable,
  includeConceptIdSetDescendant = F)
}
\description{
CreateCohort
Create cohort of interest by using condition concept Ids.
}
